requirements.txt
finnhub-python==2.4.20
python-dotenv==1.0.0
tabulate==0.9.0
requests>=2.22.0
pandas>=2.0.0
akshare>=1.10.0
